requirements.txt
django>=4.2,<4.3
djangorestframework
python-decouple
